extract outline
